chore: reorganize structure a bit
