+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
-*.js
-*.d.ts
+build
+.vscode
